+++ msmbuilder/featurizer/featurizer.py
@@ -892,26 +892,31 @@ class KappaAngleFeaturizer(Featurizer):
     """Featurizer to extract kappa angles.
 
     The kappa angle of residue `i` is the angle formed by the three CA atoms
-    of residues `i-2`, `i` and `i+2`. This featurizer extracts the
-    `n_residues - 4` kappa angles of each frame in a trajectory.
+    of residues `i-offset`, `i` and `i+offset`. This featurizer extracts the
+    `n_residues - 2*offset` kappa angles of each frame in a trajectory.
 
     Parameters
     ----------
     cos : bool
         Compute the cosine of the angle instead of the angle itself.
+    offset : int
+        Offset to use for when calculating the features. Defaults to 2.
+        I.e it will calculate the angles between i-2, i and i+2 CA
     """
 
-    def __init__(self, cos=True):
+    def __init__(self, cos=True, offset=2):
         self.cos = cos
         self.atom_indices = None
+        self.offset = offset
 
     def partial_transform(self, traj):
         ca = [a.index for a in traj.top.atoms if a.name == 'CA']
-        if len(ca) < 5:
+        if len(ca) < 2*self.offset + 1:
             return np.zeros((len(traj), 0), dtype=np.float32)
 
         angle_indices = np.array(
-            [(ca[i - 2], ca[i], ca[i + 2]) for i in range(2, len(ca) - 2)])
+            [(ca[i - self.offset], ca[i],
+              ca[i + offset]) for i in range(offset, len(ca) - offset)])
         result = md.compute_angles(traj, angle_indices)
 
         if self.atom_indices is None:
@@ -919,7 +924,6 @@ class KappaAngleFeaturizer(Featurizer):
         if self.cos:
             return np.cos(result)
 
-        assert result.shape == (traj.n_frames, traj.n_residues - 4)
         return result