+++ docs/changelog.rst
@@ -13,6 +13,8 @@ New Features
 ~~~~~~~~~~~~
 - Added new featurizer ```FeatureSlicer```. ```FeatureSlicer``` can slice the ouput of
 regular featurizer objects to just the required indices.
+- Generalized ```KappaAngleFeaturizer``` to be able compute the angles between arbitrarily
+offset CA atoms.
 - Added functions to compute error bars for transition probabilities to account for
   finite sampling, and sample transition matrices from these error distributions (i.e.
   bootstrapping). Located in ```msmbuilder.msm.validation.transmat_errorbar```.